+++ modules/action-raw.php
@@ -1,7 +1,7 @@
 <?php
 register_module([
 	"name" => "Raw page source",
-	"version" => "0.6",
+	"version" => "0.7",
 	"author" => "Starbeamrainbowlabs",
 	"description" => "Adds a 'raw' action that shows you the raw source of a page.",
 	"id" => "action-raw",
@@ -25,10 +25,9 @@ register_module([
 		 */
 		add_action("raw", function() {
 			global $env;
-
-			header("x-filename: " . rawurlencode($env->page) . ".md");
+			
 			header("content-type: text/markdown");
-			exit(file_get_contents("$env->storage_prefix$env->page.md"));
+			exit(file_get_contents($env->page_filename));
 			exit();
 		});